PoetryAI/requirements.txt
langchain
python-dotenv
streamlit
huggingface_hub
langchain_community
langchain-huggingface
huggingface_hub
langchain_google_genai